% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_transform_to_spatial.R
\name{back_transform_to_spatial}
\alias{back_transform_to_spatial}
\title{Back Transform to Spatial}
\usage{
back_transform_to_spatial(preds, transformObj)
}
\arguments{
\item{preds}{A vector of predictions from the machine learning model}

\item{transformObj}{The object outputted IndData after running transform_to_ind}
}
\value{
A vector of predictions that have been transformed back to the original spatial scale
}
\description{
This function takes a vector of predictions from a machine learning model that have been transformed to an independent scale and transforms them back to the original spatial scale.
}
\details{
Title
}
