% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspatialtransform-package.R
\docType{package}
\name{rspatialtransform-package}
\alias{rspatialtransform}
\alias{rspatialtransform-package}
\title{rspatialtransform: Spatial Transformation}
\description{
rspatialtransform transforms spatially correlated to be independent and be ready to be trained on machine learning models.
}
\author{
\strong{Maintainer}: Matthew Heaton \email{first.last@example.com}

Authors:
\itemize{
  \item Drew Millane \email{first.last@example.com}
}

}
\keyword{internal}
