% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_ind.R
\name{transform_to_ind}
\alias{transform_to_ind}
\title{Transform Spatial Data to Independent Data}
\usage{
transform_to_ind(
  formula,
  trainData,
  trainLocs,
  testData,
  testLocs,
  MaternParams = NULL,
  smoothness = 1/2,
  M = 30,
  ncores = parallel::detectCores() - 5
)
}
\arguments{
\item{formula}{An object of class "formula" describing the model to be decorrelated.}

\item{trainData}{An object of class data.frame containing the training data with the response variable provided.}

\item{trainLocs}{A matrix object containing the coordinates of the training data. The dimensions should be nx2.}

\item{testData}{An object of class data.frame containing the data to be predicted.}

\item{testLocs}{A matrix object containing the coordinates of the test data. The dimensions should be nx2.}

\item{MaternParams}{A vector of two parameters: range and nugget. Range represents how fast the correlation decays with distance and nugget represents the variability in one location. The default is NULL (rng, nug) where the range and nugget parameter are estimated automatically.}

\item{smoothness}{The smoothness parameter, which controls the smoothness of the function. The default is 1/2 which results in an exponential kernel.}

\item{M}{The number of neighbors to consider when creating a correlation matrix for each individual observation. The default is 30.}

\item{ncores}{The number of cores to parallelize the decorrelation process.}
}
\value{
list
}
\description{
This function takes a formula, training data, and test data and transforms the data to an independent scale. The function returns a list containing the transformed training data, transformed test data, range, nugget, number of neighbors, and back transformation information.
}
\details{
Title
}
